requirements.txt
Django
pytz
requests
djangorestframework
django-filter
django-cors-headers
psycopg2-binary
python-dotenv
drf-extra-fields
django-rest-passwordreset
PyYAML
ujson
gunicorn
redis
celery
sentry-sdk[django]
drf-spectacular
social-auth-app-django
djangorestframework_simplejwt
django-cachalot
django-redis
django-baton
Pillow
django-imagekit

